refactor(CurrencyUtil): add id to construct dto refactor(CurrencyJdbcDAO, ExchangeRateJdbcDAO): use try-with-resources for PreparedStatement refactor(CurrencyDAO, ExchangeRateDAO): minor corrections

+++ src/main/java/dao/ExchangeRateDAO.java
@@ -5,9 +5,9 @@ import model.ExchangeRate;
 import java.util.List;
 
 public interface ExchangeRateDAO {
-    void save(String baseCurrenciesCode, String targetCurrenciesCode, double rate);
+    void save(String baseCurrenciesCode, String targetCurrenciesCode, float rate);
 
-    void update(String baseCurrenciesCode, String targetCurrenciesCode, double rate);
+    void update(String baseCurrenciesCode, String targetCurrenciesCode, float rate);
 
     ExchangeRate get(String baseCurrenciesCode, String targetCurrenciesCode);
